+++ mayan/apps/documents/serializers/document_type_serializers.py
@@ -31,13 +31,23 @@ class DocumentTypeQuickLabelSerializer(serializers.ModelSerializer):
 
 
 class DocumentTypeSerializer(serializers.HyperlinkedModelSerializer):
+    documents_url = serializers.HyperlinkedIdentityField(
+        label=_(message='Documents URL'), lookup_url_kwarg='document_type_id',
+        view_name='rest_api:documenttype-document-list'
+    )
     quick_label_list_url = serializers.HyperlinkedIdentityField(
-        label=_(message='Quick label list URL'), lookup_url_kwarg='document_type_id',
+        label=_(message='Quick label list URL'),
+        lookup_url_kwarg='document_type_id',
         view_name='rest_api:documenttype-quicklabel-list'
     )
 
     class Meta:
         extra_kwargs = {
+            'documents_url': {
+                'label': _(message='Documents URL'),
+                'lookup_url_kwarg': 'document_type_id',
+                'view_name': 'rest_api:documenttype-document-list'
+            },
             'url': {
                 'label': _(message='URL'),
                 'lookup_url_kwarg': 'document_type_id',
@@ -47,7 +57,8 @@ class DocumentTypeSerializer(serializers.HyperlinkedModelSerializer):
         fields = (
             'delete_time_period', 'delete_time_unit',
             'document_stub_expiration_interval',
-            'document_stub_pruning_enabled', 'filename_generator_backend',
+            'document_stub_pruning_enabled', 'documents_url',
+            'filename_generator_backend',
             'filename_generator_backend_arguments', 'id', 'label',
             'quick_label_list_url', 'trash_time_period', 'trash_time_unit',
             'url'